resources/requirements.txt
paho-mqtt>=2.0.0
jeedomdaemon~=1.2.9